fix: Externalize cache updates to object
