+++ package.json
@@ -19,7 +19,7 @@
     "@aerogear/cordova-plugin-aerogear-metrics": "2.3.0",
     "@aerogear/push": "2.3.0",
     "@aerogear/security": "2.3.0",
-    "@aerogear/voyager-client": "2.3.0",
+    "@aerogear/voyager-client": "2.3.1-dev.1",
     "@angular/common": "7.2.7",
     "@angular/core": "7.2.7",
     "@angular/forms": "7.2.7",
